great overfhaul on how analysis endpoints are structured

+++ backend/core/app/models/ensemble.py
@@ -1,10 +1,11 @@
 from http.client import HTTPResponse
-from ..utils import STATUS, get_container_host
+import json
+from ..utils import STATUS, create_response_error, create_response_message, deregister_container_from_ensemble, get_container_host, parse_response_for_triggered_analysis
 from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship, Session
 from .ensemble_ids import EnsembleIds, get_ensemble_ids_by_ids
 from ..database import Base
-from .ids_container import IdsContainer
+from .ids_container import IdsContainer, update_container_status
 from ..validation.models import EnsembleUpdate
 import httpx 
 
@@ -33,8 +34,10 @@ class Ensemble(Base):
         endpoint = f"/configure/ensemble/add/{self.id}"
         async with httpx.AsyncClient() as client:
                 response: HTTPResponse = await client.post(container_url+endpoint)
-        db.add(ensemble_ids)
-        db.commit()
+        if response.status_code == 200:
+            db.add(ensemble_ids)
+            db.commit()
+        return response
     
     async def remove_container(self, container_id: int, db: Session):
         from .ids_container import IdsContainer
@@ -42,14 +45,13 @@ class Ensemble(Base):
         ensemble_ids = get_ensemble_ids_by_ids(self.id, container_id, db)
 
         container: IdsContainer = db.query(IdsContainer).filter(IdsContainer.id == container_id).first() 
-        host = get_container_host(container)
-        container_url = f"http://{host}:{container.port}"
-        endpoint = f"/configure/ensemble/remove"
-        async with httpx.AsyncClient() as client:
-                response: HTTPResponse = await client.post(container_url+endpoint)
+        response = deregister_container_from_ensemble(container)
 
-        db.delete(ensemble_ids)
-        db.commit()
+        if response.status_code == 200:
+            db.delete(ensemble_ids)
+            db.commit()
+
+        return response
 
     def get_enssemble_ids(self, db: Session):
         return db.query(EnsembleIds).filter(EnsembleIds.ensemble_id == self.id).all()
@@ -60,8 +62,59 @@ class Ensemble(Base):
         id_list = [e_ids.ids_container_id for e_ids in ensemble_ids]
         containers: list[IdsContainer] = db.query(IdsContainer).filter(IdsContainer.id.in_(id_list)).all()
         return containers
+    
+    async def start_static_analysis(self, static_analysis_data, dataset, db):
+        from .ids_container import IdsContainer
+        containers: list[IdsContainer] = self.get_containers(db)
+        responses = []
 
-def get_all_ids_ensembles(db: Session):
+        for container in containers:
+            form_data= {
+                "container_id": (None, str(container.id), "application/json"),
+                "ensemble_id": (None, str(self.id), "application/json"),
+                "file": (dataset.name, dataset.configuration, "application/octet-stream"),
+            }    
+            response: HTTPResponse = await container.start_static_analysis(form_data)
+            response = await parse_response_for_triggered_analysis(response, container, db, "static", self.id)
+            
+            if response.status_code == 200:
+                await update_container_status(STATUS.ACTIVE.value, container, db)
+            
+            responses.append(response)
+        return responses
+    
+    async def start_network_analysis(self, network_analysis_data, db):
+        from .ids_container import IdsContainer
+        containers: list[IdsContainer] = self.get_containers(db)
+        responses = []
+    
+        for container in containers:
+            data = json.dumps(network_analysis_data.__dict__)
+            response: HTTPResponse = await container.start_network_analysis(data)
+            response = await parse_response_for_triggered_analysis(response, container, db, "network", self.id)
+            # set container status to active/idle afterwards before
+            if response.status_code == 200:
+                await update_container_status(STATUS.ACTIVE.value, container, db)
+                
+            responses.append(response)  
+        return responses
+
+    async def stop_analysis(self):
+        containers: list[IdsContainer] = self.get_containers(db)
+
+        responses = []
+
+        for container in containers:
+            response: HTTPResponse = await container.stop_analysis()
+            if response.status_code == 200:
+                message= f"Analysis for container {container.id} successfully stopped"
+                responses.append(create_response_message(message, 200))
+            else:
+                message=f"Analysis for container {container.id} could not be stopped"
+                responses.append(create_response_error(message, 500)) 
+        return responses
+    
+def get_all_ensembles(db: Session):
     return db.query(Ensemble).all()
 
 # TODO: make all db actions asynchronous
@@ -91,12 +144,15 @@ async def update_ensemble(ensemble: EnsembleUpdate, db: Session):
     added_containers = list(filter(lambda x: x not in former_containers, new_containers))
     removed_containers = list(filter(lambda x: x not in new_containers, former_containers))
 
-    for container_id in removed_containers:
-        await ensemble_db.remove_container(container_id, db)
-    for container_id in added_containers:
-        await ensemble_db.add_container(container_id, db)
+    responses = []
 
-    return {"message": "Successfully updated Ensemble"}
+    for container_id in removed_containers:
+        res = await ensemble_db.remove_container(container_id, db)
+        responses.append(res)
+    for container_id in added_containers:
+        res = await ensemble_db.add_container(container_id, db)
+        responses.append(res)
+    return responses
 
 
 async def update_ensemble_status(status: STATUS, ensemble: Ensemble, db: Session):
